- [Seller] 제품판매 등록 / 수정 / 삭제 - [Buyer / Seller] 주문관리 목록 확인 / Buyer [의뢰완료 / 의뢰취소] 기능구현

+++ src/main/java/project/TadeM/product/entity/product.java
@@ -26,7 +26,7 @@ public class product {
 
 	String name; // 제품 이름
 	String contents; // 제품 내용
-	int price; // 제품 가격
+	Long price; // 제품 가격
 	LocalDateTime regDt; // 제품 등록일
 	String period; // 작업기간
 }
